refactor: clean up directories
