+++ src/app/auth/(docs)/privacy-policy/page.tsx
