+++ src/app/auth/(docs)/terms-of-service/page.tsx
