--- src/app/globals.css
@@ -1,4 +1,0 @@
-@import './styles/colors.css';
-@import './styles/boilerplate.css';
-@import './styles/typography.css';
-@import './styles/common.css';
