+++ src/app/providers/ReduxProvider.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Provider } from 'react-redux';
-import { store } from '@/services/store'
+import { store } from '@/app/store/store'
 
 
 export function ReduxProvider({ children }: { children: React.ReactNode }) {
