+++ src/features/auth/hooks/useAuth.ts
@@ -1,4 +1,4 @@
-import { AuthContext } from '@/shared/providers/AuthProvider'
+import { AuthContext } from '@/app/providers/AuthProvider'
 import { useContext } from 'react'
 
 export const useAuth = () => {
