+++ src/shared/ui/header/Header.tsx
@@ -5,8 +5,8 @@ import s from './Header.module.scss'
 import { SelectBox, SelectItem } from '@/shared/ui/selectBox/SelectBox'
 import { Button } from '@/shared/ui/button'
 import { IoNotificationsOutline } from 'react-icons/io5'
-import flagUnitedKingdom from '@/assets/images/flagUnitedKingdom.png'
-import flagRussia from '@/assets/images/flagRussia.png'
+import flagUnitedKingdom from '@/shared/ui/header/assets/flagUnitedKingdom.png'
+import flagRussia from '@/shared/ui/header/assets/flagRussia.png'
 import Image, { StaticImageData } from 'next/image'
 import Link from 'next/link'
 import { useAuth } from '@/features/auth/hooks/useAuth'
