+++ src/shared/ui/header/assets/flagRussia.png
