+++ src/shared/ui/header/assets/flagUnitedKingdom.png
