+++ src/styles/globals.css
@@ -1,0 +1,4 @@
+@import 'colors.css';
+@import 'boilerplate.css';
+@import 'typography.css';
+@import 'common.css';
